moved question widget to separate file
